Downloads/Rl_unidepth_final/requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pyyaml==6.0
pillow==10.0.0
matplotlib==3.7.2
scipy==1.11.1
tqdm==4.65.0
tensorboard==2.13.0

# Optional dependencies
# fvcore==0.1.1.post20230915  # For FLOPs counting
# wandb==0.15.8  # For W&B logging
